feat(offers): link MealPlanTemplate to offers

+++ backend/src/main/java/org/kmurygin/healthycarbs/offers/OfferService.java
@@ -2,17 +2,18 @@ package org.kmurygin.healthycarbs.offers;
 
 import lombok.RequiredArgsConstructor;
 import org.kmurygin.healthycarbs.exception.ResourceNotFoundException;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
 @RequiredArgsConstructor
-@RestController
+@Service
 public class OfferService {
 
     private final OfferRepository offerRepository;
     private final OfferMapper offerMapper;
+    private final MealPlanTemplateRepository templateRepository;
 
     public List<Offer> findAll() {
         return offerRepository.findAll();
@@ -23,7 +24,18 @@ public class OfferService {
                 .orElseThrow(() -> new ResourceNotFoundException("Offer", "id", id));
     }
 
+    @Transactional
     public Offer create(Offer offer) {
+        return offerRepository.save(offer);
+    }
+
+    @Transactional
+    public Offer create(Offer offer, Long mealPlanTemplateId) {
+        MealPlanTemplate template = templateRepository.findById(mealPlanTemplateId)
+                .orElseThrow(() -> new ResourceNotFoundException(
+                        "MealPlanTemplate", "id", mealPlanTemplateId)
+                );
+        offer.setMealPlanTemplate(template);
         return offerRepository.save(offer);
     }
 
@@ -36,7 +48,26 @@ public class OfferService {
         return offerRepository.save(offer);
     }
 
+    @Transactional
+    public Offer update(Long id, Offer updatedOffer, Long mealPlanTemplateId) {
+        Offer offer = offerRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Offer", "id", id));
+
+        offerMapper.updateFromEntity(updatedOffer, offer);
+        MealPlanTemplate template = templateRepository.findById(mealPlanTemplateId)
+                .orElseThrow(() -> new ResourceNotFoundException(
+                        "MealPlanTemplate", "id", mealPlanTemplateId)
+                );
+        offer.setMealPlanTemplate(template);
+
+        return offerRepository.save(offer);
+    }
+
+    @Transactional
     public void deleteById(Long id) {
+        if (!offerRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Offer", "id", id);
+        }
         offerRepository.deleteById(id);
     }
 }
